+++ demos/chinook/src/main/java/module-info.java
@@ -24,7 +24,7 @@ module is.codion.framework.demos.chinook {
 	requires is.codion.framework.db.local;
 	requires is.codion.framework.db.http;
 	requires is.codion.framework.db.rmi;
-	requires is.codion.swing.common.ui.tools;
+	requires is.codion.swing.common.loadtest.ui;
 	requires is.codion.swing.framework.ui;
 	requires is.codion.plugin.jasperreports;
 	requires is.codion.plugin.imagepanel;