+++ demos/employees/src/main/java/module-info.java
@@ -26,7 +26,7 @@ module is.codion.framework.demos.employees {
 	requires is.codion.framework.db.rmi;
 	requires is.codion.framework.domain.test;
 	requires is.codion.framework.json.domain;
-	requires is.codion.swing.common.ui.tools;
+	requires is.codion.swing.common.loadtest.ui;
 	requires is.codion.swing.framework.ui;
 	requires is.codion.framework.server;
 	requires is.codion.plugin.jasperreports;