+++ swing/common-loadtest-ui/src/main/java/module-info.java
@@ -17,17 +17,15 @@
  * Copyright (c) 2023 - 2024, Björn Darri Sigurðsson.
  */
 /**
- * Petstore demo.
+ * Common Swing loadtest UI.<br>
+ * <br>
+ * {@link is.codion.swing.common.loadtest.ui.LoadTestPanel}<br>
  */
-module is.codion.framework.demos.petstore {
-	requires is.codion.swing.common.loadtest.ui;
-	requires is.codion.swing.framework.ui;
+module is.codion.swing.common.loadtest.ui {
+	requires org.jfree.jfreechart;
+	requires com.formdev.flatlaf.intellijthemes;
+	requires transitive is.codion.swing.common.loadtest.model;
+	requires transitive is.codion.swing.common.ui;
 
-	exports is.codion.framework.demos.petstore.model
-					to is.codion.swing.framework.model, is.codion.swing.framework.ui;
-	exports is.codion.framework.demos.petstore.ui
-					to is.codion.swing.framework.ui;
-
-	provides is.codion.framework.domain.Domain
-					with is.codion.framework.demos.petstore.domain.Petstore;
+	exports is.codion.swing.common.loadtest.ui;
 }